Validate message from Front
The whole point of this isolated storage design is to protect against XSS
If Front, the REACT app, is poisoned with XSS, we still want protection
So we should validate all messages from Front

Currently, validation checks that the label property is a string
Without this, an attacker could pass an evil object
The evil object could have an evil toString() method
The method might return a valid string
But it could also do bad stuff e.g. exfiltrate local storage
So we check that the type is indeed an actual string
We do this before any further processing

Right now, a non-string doesn't seem to pose a threat
But in future, we might assume this is a string and add it to a string
E.g. we might interpolate it into a log message
Such action will cause toString() to fire
Which would propagate the XSS from Front to Hub